:hotfix: 인터페이스 수정 및 community 핫픽스

+++ src/main/java/com/aqh/board/dao/BoardDAO.java
@@ -20,7 +20,7 @@ public interface BoardDAO {
 
 	public int getBoardTotal(Criteria criteria);
 	
-	public long getReplyTotal(long bno);
+	public int getReplyTotal(long bno);
 
 	public BoardDTO findByBoardNumber(long bno);
